requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
requests>=2.25.1
vosk==0.3.44
pydub>=0.25.1
ffmpeg-python>=0.2.0
websockets>=10.0
openai-whisper>=20231117
torch>=2.1.0
numpy>=1.24.0
